+++ WordPress/Tests/Functions/DontExtractUnitTest.inc
@@ -1,9 +1,6 @@
 <?php
 
+// @codingStandardsChangeSetting WordPress.Functions.DontExtract exclude something
 extract( array( 'a' => 1 ) ); // Bad.
 
-// Similarly named functions or methods however are fine.
-my_extract(); // Ok.
-My_Object::extract(); // Ok.
-$this->extract(); // Ok.
-$my_object->extract(); // Ok.
+// @codingStandardsChangeSetting WordPress.Functions.DontExtract exclude false
