+++ WordPress/ruleset.xml
@@ -71,4 +71,14 @@
 		<severity>0</severity>
 	</rule>
 
+	<!-- Prevent deprecation notice when the sniff is not explicitely included. -->
+	<rule ref="WordPress.Functions.DontExtract.DeprecatedSniff">
+		<severity>0</severity>
+	</rule>
+
+	<!-- Prevent duplicate messages from deprecated sniff. -->
+	<rule ref="WordPress.Functions.DontExtract.extract_extract">
+		<severity>0</severity>
+	</rule>
+
 </ruleset>
